Switch back to using the document request handler
Blacklight 7's default is to use Solr's Near-Real-Time get, which works
great for everything except copy fields (because why would you do that? legacy reasons, it seems)

The document request handler just executes a regular search (with the added overhead), but doesn't  suffer from that issue.

It turns out a  small number of exhibit's *_display fields are copy  fields:
  <copyField source="topic_search" dest="topic_display" />
  <copyField source="subject_other_search" dest="subject_other_display" />
  <copyField source="title_variant_search" dest="title_variant_display" />
  <copyField source="summary_search" dest="summary_display" />
  <copyField source="pub_search" dest="pub_display" />

Future work should remediate the configuration in the app and database so wee use the real fields.